+++ xoutil/functools.py
@@ -24,9 +24,9 @@
 # Created on Feb 22, 2012
 
 '''
-The `functools` module is for higher-order functions: functions that act on or
-return other functions. In general, any callable object can be treated as a
-function for the purposes of this module.
+Extensions to the `functools` module from the Python's standard library.
+
+You may use this module as drop-in replacement of `functools`.
 '''
 
 from __future__ import (division as _py3_division,
@@ -34,81 +34,114 @@ from __future__ import (division as _py3_division,
                         unicode_literals as _py3_unicode,
                         absolute_import as _py3_abs_imports)
 
-__docstring_format__ = 'rst'
-__author__ = 'manu'
-
 
 from functools import *
-
-WRAPPER_ASSIGNMENTS = ('__module__', '__name__')
-WRAPPER_UPDATES = ('__dict__',)
-_update_wrapper = update_wrapper
-
-def update_wrapper(wrapper,
-                   wrapped,
-                   assigned=WRAPPER_ASSIGNMENTS,
-                   updated=WRAPPER_UPDATES,
-                   add_signature=False):
-    """Update a wrapper function to look like the wrapped function
-
-       wrapper is the function to be updated
-       wrapped is the original function
-       assigned is a tuple naming the attributes assigned directly
-       from the wrapped function to the wrapper function (defaults to
-       functools.WRAPPER_ASSIGNMENTS)
-       updated is a tuple naming the attributes of the wrapper that
-       are updated with the corresponding attribute from the wrapped
-       function (defaults to functools.WRAPPER_UPDATES)
-
-       Fixed on xoutil:
-
-       Since most decorators use the *args, **kwargs idiom for arguments it is
-       introduced the `add_signature` argument to prepend the signature to the
-       documentation.
-    """
-    from inspect import getargspec
-    result = _update_wrapper(wrapper, wrapped, assigned, updated)
-    doc = getattr(wrapped, '__doc__')
-    if add_signature:
-        module = getattr(wrapped, '__module__')
-        name = getattr(wrapped, '__name__')
-        spec = getargspec(wrapped)
-        rsign = []
-        nargs_without_default_values = len(spec.args) - len(spec.defaults or [])
-        rsign.extend(spec.args[:nargs_without_default_values])
-        if spec.defaults:
-            rsign.extend('{arg}={val!r}'.format(arg=spec.args[nargs_without_default_values + i],
-                                                val=spec.defaults[i])
-                            for i in range(len(spec.defaults)))
-        if spec.varargs:
-            rsign.append('*' + spec.varargs)
-        if spec.keywords:
-            rsign.append('**' + spec.keywords)
-        signature = ', '.join(rsign)
-        doc = "{module}:{name}({signature})\n\n{doc}".format(**locals())
-    setattr(wrapper, '__doc__', doc)
-    return result
+from xoutil.compat import py32
 
 
-def wraps(wrapped,
-          assigned=WRAPPER_ASSIGNMENTS,
-          updated=WRAPPER_UPDATES,
-          add_signature=False):
-    """Decorator factory to apply update_wrapper() to a wrapper function
+if not py32:
+    from threading import Lock
+    from xoutil.collections import _CacheInfo, OrderedDict
 
-       Returns a decorator that invokes update_wrapper() with the decorated
-       function as the wrapper argument and the arguments to wraps() as the
-       remaining arguments. Default arguments are as for update_wrapper().
-       This is a convenience function to simplify applying partial() to
-       update_wrapper().
+    # Back-ported lru_cache from py32. But take note that if running with at
+    # least py32 we will use Python's version, so don't mess with internals.
+    def lru_cache(maxsize=100):
+        """Least-recently-used cache decorator.
 
-       Fixed on xoutil:
+        If *maxsize* is set to None, the LRU features are disabled and the
+        cache can grow without bound.
 
-       Since most decorators use the *args, **kwargs idiom for arguments it is
-       introduced the `add_signature` argument to prepend the signature to the
-       documentation.
-    """
-    return partial(update_wrapper, wrapped=wrapped,
-                   assigned=assigned, updated=updated,
-                   add_signature=add_signature)
+        Arguments to the cached function must be hashable.
 
+        View the cache statistics named tuple (hits, misses, maxsize,
+        currsize) with f.cache_info().  Clear the cache and statistics with
+        f.cache_clear(). Access the underlying function with f.__wrapped__.
+
+        See:  http://en.wikipedia.org/wiki/Cache_algorithms#Least_Recently_Used
+
+        """
+        # Users should only access the lru_cache through its public API:
+        #       cache_info, cache_clear, and f.__wrapped__
+        # The internals of the lru_cache are encapsulated for thread safety and
+        # to allow the implementation to change (including a possible C version).
+
+        def decorating_function(user_function,
+                    tuple=tuple, sorted=sorted, len=len, KeyError=KeyError):
+
+            _cache_info = [0, 0]
+            kwd_mark = (object(),)          # separates positional and keyword args
+            lock = Lock()                   # needed because OrderedDict isn't threadsafe
+
+            if maxsize is None:
+                cache = dict()              # simple cache without ordering or size limit
+
+                @wraps(user_function)
+                def wrapper(*args, **kwds):
+                    #~ nonlocal hits, misses
+                    hits, misses = _cache_info
+                    key = args
+                    if kwds:
+                        key += kwd_mark + tuple(sorted(kwds.items()))
+                    try:
+                        result = cache[key]
+                        hits += 1
+                        _cache_info[0] = hits
+                        _cache_info[1] = misses
+                        return result
+                    except KeyError:
+                        pass
+                    result = user_function(*args, **kwds)
+                    cache[key] = result
+                    misses += 1
+                    _cache_info[0] = hits
+                    _cache_info[1] = misses
+                    return result
+            else:
+                cache = OrderedDict()           # ordered least recent to most recent
+                cache_popitem = cache.popitem
+                cache_renew = cache.move_to_end
+
+                @wraps(user_function)
+                def wrapper(*args, **kwds):
+                    #~ nonlocal hits, misses
+                    hits, misses = _cache_info
+                    key = args
+                    if kwds:
+                        key += kwd_mark + tuple(sorted(kwds.items()))
+                    with lock:
+                        try:
+                            result = cache[key]
+                            cache_renew(key)    # record recent use of this key
+                            hits += 1
+                            _cache_info[0] = hits
+                            _cache_info[1] = misses
+                            return result
+                        except KeyError:
+                            pass
+                    result = user_function(*args, **kwds)
+                    with lock:
+                        cache[key] = result     # record recent use of this key
+                        misses += 1
+                        if len(cache) > maxsize:
+                            cache_popitem(0)    # purge least recently used cache entry
+                    _cache_info[0] = hits
+                    _cache_info[1] = misses
+                    return result
+
+            def cache_info():
+                """Report cache statistics"""
+                with lock:
+                    return _CacheInfo(_cache_info[0], _cache_info[1], maxsize, len(cache))
+
+            def cache_clear():
+                """Clear the cache and cache statistics"""
+                #~ nonlocal hits, misses
+                with lock:
+                    cache.clear()
+                    _cache_info[0] = _cache_info[1] = 0
+
+            wrapper.cache_info = cache_info
+            wrapper.cache_clear = cache_clear
+            return wrapper
+
+        return decorating_function