+++ xoutil/tests/test_collections.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
+# -*- encoding: utf-8 -*-
 #----------------------------------------------------------------------
-# xoutil.release
+# xoutil.tests.test_collections
 #----------------------------------------------------------------------
 # Copyright (c) 2012 Merchise Autrement
 # All rights reserved.
@@ -21,12 +21,31 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
 # MA 02110-1301, USA.
 #
-# Created on 2012-04-01
+# Created on Jul 3, 2012
 
 
 from __future__ import (division as _py3_division,
                         print_function as _py3_print,
                         unicode_literals as _py3_unicode,
-                        absolute_import as _py3_abs_imports)
+                        absolute_import as _absolute_import)
 
-VERSION = '1.1.0'
+import unittest
+
+from xoutil.collections import defaultdict
+
+__docstring_format__ = 'rst'
+__author__ = 'manu'
+
+
+
+class TestCollections(unittest.TestCase):
+    def test_defaultdict(self):
+        d = defaultdict(lambda key, d: 'a')
+        self.assertEqual('a', d['abc'])
+        d['abc'] = 1
+        self.assertEqual(1, d['abc'])
+
+
+if __name__ == "__main__":
+    #import sys;sys.argv = ['', 'Test.testName']
+    unittest.main(verbosity=2)